refactor: extract validation logic to new file
