+++ internal/controller/mcpserver_controller.go
@@ -63,13 +63,6 @@ const (
 // MCPClientVersion is the version sent during MCP handshake. Bump with releases.
 var MCPClientVersion = "v0.1.0"
 
-// ValidationError represents a permanent configuration validation error.
-// These errors indicate the MCPServer configuration is invalid and should not be retried.
-type ValidationError struct {
-	Reason  string
-	Message string
-}
-
 func (e *ValidationError) Error() string {
 	return e.Message
 }
@@ -584,209 +577,6 @@ func (r *MCPServerReconciler) applyStatus(
 		client.FieldOwner(fieldManager),
 		client.ForceOwnership,
 	)
-}
-
-// classifyAPIError classifies a Kubernetes API error as either a permanent ValidationError
-// or a transient error that should be retried.
-// NotFound and BadRequest are permanent — NotFound is safe to treat as permanent because the
-// controller watches ConfigMaps/Secrets and will re-reconcile when the missing resource is created.
-// All other errors (Forbidden, Unauthorized, 500, 503, 429, timeouts...) are transient.
-func classifyAPIError(resourceDesc string, namespace string, err error) error {
-	if apierrors.IsNotFound(err) {
-		return &ValidationError{
-			Reason:  ReasonInvalid,
-			Message: fmt.Sprintf("%s not found in namespace '%s'", resourceDesc, namespace),
-		}
-	}
-	if apierrors.IsBadRequest(err) {
-		return &ValidationError{
-			Reason:  ReasonInvalid,
-			Message: fmt.Sprintf("Invalid %s reference: %v", resourceDesc, err),
-		}
-	}
-	return fmt.Errorf("transient error validating %s: %w", resourceDesc, err)
-}
-
-// validateReferencedConfigMap returns a permanent ValidationError on NotFound/BadRequest,
-// or a wrapped transient error for other API failures.
-func (r *MCPServerReconciler) validateReferencedConfigMap(
-	ctx context.Context,
-	namespace, name, resourceDesc string,
-) error {
-	cm := &corev1.ConfigMap{}
-	if err := r.Get(ctx, client.ObjectKey{Name: name, Namespace: namespace}, cm); err != nil {
-		return classifyAPIError(resourceDesc, namespace, err)
-	}
-	return nil
-}
-
-// validateReferencedSecret returns a permanent ValidationError on NotFound/BadRequest,
-// or a wrapped transient error for other API failures.
-func (r *MCPServerReconciler) validateReferencedSecret(
-	ctx context.Context,
-	namespace, name, resourceDesc string,
-) error {
-	secret := &corev1.Secret{}
-	if err := r.Get(ctx, client.ObjectKey{Name: name, Namespace: namespace}, secret); err != nil {
-		return classifyAPIError(resourceDesc, namespace, err)
-	}
-	return nil
-}
-
-// validateStorageMount validates a single storage mount configuration.
-// Returns ValidationError for permanent configuration errors, wrapped error for transient errors, or nil for success.
-func (r *MCPServerReconciler) validateStorageMount(
-	ctx context.Context,
-	mcpServer *mcpv1alpha1.MCPServer,
-	storage mcpv1alpha1.StorageMount,
-	index int,
-) error {
-	switch storage.Source.Type {
-	case mcpv1alpha1.StorageTypeConfigMap:
-		if storage.Source.ConfigMap == nil {
-			return &ValidationError{
-				Reason:  ReasonInvalid,
-				Message: fmt.Sprintf("ConfigMap must be set for storage mount at index %d", index),
-			}
-		}
-		if storage.Source.ConfigMap.Name == "" {
-			return &ValidationError{
-				Reason:  ReasonInvalid,
-				Message: fmt.Sprintf("ConfigMap name must not be empty for storage mount at index %d", index),
-			}
-		}
-		// Skip validation if optional
-		if storage.Source.ConfigMap.Optional != nil && *storage.Source.ConfigMap.Optional {
-			return nil
-		}
-		return r.validateReferencedConfigMap(ctx, mcpServer.Namespace, storage.Source.ConfigMap.Name,
-			fmt.Sprintf("ConfigMap '%s'", storage.Source.ConfigMap.Name))
-
-	case mcpv1alpha1.StorageTypeSecret:
-		if storage.Source.Secret == nil {
-			return &ValidationError{
-				Reason:  ReasonInvalid,
-				Message: fmt.Sprintf("Secret must be set for storage mount at index %d", index),
-			}
-		}
-		if storage.Source.Secret.SecretName == "" {
-			return &ValidationError{
-				Reason:  ReasonInvalid,
-				Message: fmt.Sprintf("Secret name must not be empty for storage mount at index %d", index),
-			}
-		}
-		// Skip validation if optional
-		if storage.Source.Secret.Optional != nil && *storage.Source.Secret.Optional {
-			return nil
-		}
-		return r.validateReferencedSecret(ctx, mcpServer.Namespace, storage.Source.Secret.SecretName,
-			fmt.Sprintf("Secret '%s'", storage.Source.Secret.SecretName))
-
-	case mcpv1alpha1.StorageTypeEmptyDir:
-		// Validate EmptyDir configuration is present
-		if storage.Source.EmptyDir == nil {
-			return &ValidationError{
-				Reason:  ReasonInvalid,
-				Message: fmt.Sprintf("EmptyDir must be set for storage mount at index %d", index),
-			}
-		}
-
-	default:
-		// Unknown/unsupported storage type
-		return &ValidationError{
-			Reason:  ReasonInvalid,
-			Message: fmt.Sprintf("Unsupported storage type '%s' at index %d", storage.Source.Type, index),
-		}
-	}
-	return nil
-}
-
-// validateEnvFrom validates a single envFrom configuration.
-// Returns ValidationError for permanent configuration errors, wrapped error for transient errors, or nil for success.
-func (r *MCPServerReconciler) validateEnvFrom(
-	ctx context.Context,
-	mcpServer *mcpv1alpha1.MCPServer,
-	envFrom corev1.EnvFromSource,
-	index int,
-) error {
-	if ref := envFrom.ConfigMapRef; ref != nil {
-		if ref.Optional == nil || !*ref.Optional {
-			if err := r.validateReferencedConfigMap(ctx, mcpServer.Namespace, ref.Name,
-				fmt.Sprintf("ConfigMap '%s' (envFrom index %d)", ref.Name, index)); err != nil {
-				return err
-			}
-		}
-	}
-	if ref := envFrom.SecretRef; ref != nil {
-		if ref.Optional == nil || !*ref.Optional {
-			if err := r.validateReferencedSecret(ctx, mcpServer.Namespace, ref.Name,
-				fmt.Sprintf("Secret '%s' (envFrom index %d)", ref.Name, index)); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// validateEnvValueFrom validates a single env var's valueFrom configuration.
-// Returns ValidationError for permanent configuration errors, wrapped error for transient errors, or nil for success.
-func (r *MCPServerReconciler) validateEnvValueFrom(
-	ctx context.Context,
-	mcpServer *mcpv1alpha1.MCPServer,
-	env corev1.EnvVar,
-	index int,
-) error {
-	if env.ValueFrom == nil {
-		return nil
-	}
-	if ref := env.ValueFrom.ConfigMapKeyRef; ref != nil {
-		if ref.Optional == nil || !*ref.Optional {
-			if err := r.validateReferencedConfigMap(ctx, mcpServer.Namespace, ref.Name,
-				fmt.Sprintf("ConfigMap '%s' referenced by env var '%s' (env index %d)", ref.Name, env.Name, index)); err != nil {
-				return err
-			}
-		}
-	}
-	if ref := env.ValueFrom.SecretKeyRef; ref != nil {
-		if ref.Optional == nil || !*ref.Optional {
-			if err := r.validateReferencedSecret(ctx, mcpServer.Namespace, ref.Name,
-				fmt.Sprintf("Secret '%s' referenced by env var '%s' (env index %d)", ref.Name, env.Name, index)); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// validateConfig validates the MCPServer configuration.
-// Returns ValidationError for permanent configuration errors, wrapped error for transient errors, or nil for success.
-func (r *MCPServerReconciler) validateConfig(
-	ctx context.Context,
-	mcpServer *mcpv1alpha1.MCPServer,
-) error {
-	// Validate storage mounts
-	for i, storage := range mcpServer.Spec.Config.Storage {
-		if err := r.validateStorageMount(ctx, mcpServer, storage, i); err != nil {
-			return err
-		}
-	}
-
-	// Validate envFrom references
-	for i, envFrom := range mcpServer.Spec.Config.EnvFrom {
-		if err := r.validateEnvFrom(ctx, mcpServer, envFrom, i); err != nil {
-			return err
-		}
-	}
-
-	// Validate env valueFrom references
-	for i, env := range mcpServer.Spec.Config.Env {
-		if err := r.validateEnvValueFrom(ctx, mcpServer, env, i); err != nil {
-			return err
-		}
-	}
-
-	// All validation passed
-	return nil
 }
 
 // extractConfigMapNames is an index extractor that returns all ConfigMap names
